Move gltf into new formats package
