+++ src/main/kotlin/models/formats/MeshFormatExporter.kt
@@ -1,6 +1,6 @@
 package models.formats
 
-import models.glTF.MaterialBuffers
+import models.formats.glTF.MaterialBuffers
 import utils.ChunkWriteListener
 
 /** A buffered 3D mesh file format writer. */
