+++ src/main/kotlin/models/formats/glTF/AccessorType.kt
@@ -1,4 +1,4 @@
-package models.glTF
+package models.formats.glTF
 
 enum class AccessorType(val value: String) {
     SCALAR("SCALAR"),
