+++ src/main/kotlin/models/formats/glTF/Buffer.kt
@@ -1,4 +1,4 @@
-package models.glTF
+package models.formats.glTF
 
 class Buffer(filename: String, val byteLength: Long) {
     val uri = "$filename.bin"
