+++ src/main/kotlin/models/formats/glTF/BufferView.kt
@@ -1,4 +1,4 @@
-package models.glTF
+package models.formats.glTF
 
 class BufferView(val buffer: Int, val byteOffset: Long, val byteLength: Long) {
     val target = BufferTarget.ARRAY_BUFFER.value
