+++ src/main/kotlin/models/formats/glTF/Image.kt
@@ -1,3 +1,3 @@
-package models.glTF
+package models.formats.glTF
 
 class Image(val uri: String)
