+++ src/main/kotlin/models/formats/glTF/MaterialBuffers.kt
@@ -1,4 +1,4 @@
-package models.glTF
+package models.formats.glTF
 
 import cache.utils.ColorPalette
 import java.awt.Color
