+++ src/main/kotlin/models/formats/glTF/Mesh.kt
@@ -1,3 +1,3 @@
-package models.glTF
+package models.formats.glTF
 
 class Mesh(val primitives: ArrayList<Primitive>)
