+++ src/main/kotlin/models/formats/glTF/Node.kt
@@ -1,4 +1,4 @@
-package models.glTF
+package models.formats.glTF
 
 class Node(
     val mesh: Int
