+++ src/main/kotlin/models/formats/glTF/Primitive.kt
@@ -1,4 +1,4 @@
-package models.glTF
+package models.formats.glTF
 
 import com.fasterxml.jackson.annotation.JsonInclude
