+++ src/main/kotlin/models/formats/glTF/Scene.kt
@@ -1,4 +1,4 @@
-package models.glTF
+package models.formats.glTF
 
 class Scene(
     val nodes: ArrayList<Int>
