+++ src/main/kotlin/models/formats/glTF/Texture.kt
@@ -1,3 +1,3 @@
-package models.glTF
+package models.formats.glTF
 
 class Texture(val source: Int)
