+++ src/main/kotlin/models/formats/glTF/extensions/KHRMaterialsSpecular.kt
@@ -1,3 +1,3 @@
-package models.glTF.extensions
+package models.formats.glTF.extensions
 
 class KHRMaterialsSpecular(val specularFactor: Float, val specularColorFactor: Array<Float>)
